flask_backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dateutil==2.8.2
bcrypt==4.0.1

# ML & Behavioral Analysis
scikit-learn
numpy
pandas
scipy
joblib
